+++ kamehouse-ui/src/main/webapp/js/admin/entity-manager/kamehouse-user.js
@@ -96,7 +96,7 @@ $(document).ready(() => {
             roles.forEach((role) => {
               rolesToPrint.push(role.name);
             });
-            return JSON.stringify(rolesToPrint);
+            return kameHouse.json.stringify(rolesToPrint);
           }
         },
         {